Delete GeoPointMapActivity
